issue nobodies-collective/Humans#697: step 4 — wire controller, delete dead surface
`AccountController.ExternalLoginCallback` now calls
`ReconcileOAuthIdentityAsync` exactly once on each of the five
OAuth-success paths: existing-user sign-in success, already-authenticated
link, lockout-relink, email-match link, and new-user creation. The
controller writes no audit rows on the OAuth path — that ownership lives
in the service per the issue spec.

`DetectAndApplyRenameAsync` and `TryLinkProviderForUserEmailAsync` are
gone. The controller passes the OIDC `email_verified` boolean through
to the service via a new `claimEmailVerified` argument; `Program.cs`
adds the `MapJsonKey("email_verified", ...)` so Google's claim survives
the GoogleHandler.

Dead surface removed:

- `IUserEmailService.FindByProviderKeyAsync`, `UpdateEmailAsync`,
  `LinkAsync` and the `UserEmailProviderMatch` record.
- `IUserEmailRepository.FindAllByProviderKeyAsync` and
  `UpdateEmailAsync` (the old upsert primitive). The repository's
  `_logger` field went with `UpdateEmailAsync` — the now-empty
  constructor takes only the DbContext factory.
- `EmailProblemsService.BackfillLegacyIdentityEmailsAsync`'s
  branch that called `LinkAsync` when an external login existed. The
  legacy address is added as a plain verified row; the next OAuth
  callback's reconcile attaches the provider tag.
- Tests for the dead methods (corrupted-state / foreign-user-match
  scenarios that the new policy makes unreachable):
  `AccountControllerOAuthRenameDetectionTests` (replaced by
  `AccountControllerOAuthReconcileTests` covering the five paths),
  `UserEmailServiceProviderInvariantsTests` (entire file — pinned the
  dead `FindByProviderKeyAsync`), the `LinkAsync` / `UpdateEmailAsync`
  blocks in `UserEmailServiceTests`, and the with-external-login
  branch in `EmailProblemsServiceTests`.

Doc-comment references to `IUserEmailService.LinkAsync` on the
`GetExternalLoginsByUserIdsAsync` contracts updated to point at the
admin diagnostic + reconcile mother-of-all log (step 5 owners).
