+++ src/Humans.Application/Interfaces/Repositories/IUserRepository.cs
@@ -190,10 +190,9 @@ public interface IUserRepository : IRepository
     /// <summary>
     /// Returns every <c>AspNetUserLogins</c> <c>(LoginProvider, ProviderKey)</c>
     /// row for each of the given users, grouped by <c>UserId</c>. Users without
-    /// any external login are absent from the dictionary. Used by the legacy
-    /// email backfill to pick a <c>(Provider, ProviderKey)</c> pair so the
-    /// inserted <c>UserEmail</c> row can be linked via
-    /// <c>IUserEmailService.LinkAsync</c> in one step instead of two.
+    /// any external login are absent from the dictionary. Used by the admin
+    /// per-user emails diagnostic to show the OAuth identity store alongside
+    /// the <c>UserEmail</c> tag rows.
     /// </summary>
     Task<IReadOnlyDictionary<Guid, IReadOnlyList<(string Provider, string ProviderKey)>>>
         GetExternalLoginsByUserIdsAsync(
