+++ src/Humans.Application/Services/Profile/EmailProblemsService.cs
@@ -166,10 +166,13 @@ public sealed class EmailProblemsService : IEmailProblemsService
     public async Task<IReadOnlyList<(Guid UserId, string Email)>> BackfillLegacyIdentityEmailsAsync(
         Guid actorUserId, CancellationToken ct = default)
     {
+        // actorUserId is captured by the caller's audit row; the scanner is
+        // admin-invoked and the per-row audit lives at the controller level.
+        _ = actorUserId;
+
         var users = await _userService.GetAllUsersAsync(ct);
         var userIds = users.Select(u => u.Id).ToList();
         var emailsByUser = await _userEmailService.GetEntitiesByUserIdsAsync(userIds, ct);
-        var loginsByUser = await _userService.GetExternalLoginsByUserIdsAsync(userIds, ct);
 
         var backfilled = new List<(Guid, string)>();
 
@@ -185,24 +188,12 @@ public sealed class EmailProblemsService : IEmailProblemsService
                 && string.Equals(e.Email, legacy, StringComparison.OrdinalIgnoreCase)))
                 continue;
 
-            // Prefer LinkAsync when the user has an external login — it inserts
-            // the UserEmail row already tagged with (Provider, ProviderKey) so a
-            // subsequent OAuth sign-in finds the row by FindByProviderKeyAsync
-            // and doesn't need to re-tag. Falls back to AddVerifiedEmailAsync
-            // for users with no OAuth login (legacy magic-link-only accounts).
-            // When multiple external logins exist, the first one wins —
-            // additional providers self-heal on their respective OAuth callbacks
-            // via the rename/backfill branch in AccountController.
-            if (loginsByUser.TryGetValue(u.Id, out var logins) && logins.Count > 0)
-            {
-                var (provider, providerKey) = logins[0];
-                await _userEmailService.LinkAsync(
-                    u.Id, provider, providerKey, legacy, actorUserId: actorUserId, ct);
-            }
-            else
-            {
-                await _userEmailService.AddVerifiedEmailAsync(u.Id, legacy, ct);
-            }
+            // Issue nobodies-collective/Humans#697: write the legacy address as
+            // a plain verified row. The (Provider, ProviderKey) tag is no
+            // longer authoritative for OAuth identity (AspNetUserLogins is) —
+            // the next OAuth sign-in's reconcile finds the matching row by
+            // address and attaches the tag via TagMoved.
+            await _userEmailService.AddVerifiedEmailAsync(u.Id, legacy, ct);
             backfilled.Add((u.Id, legacy));
         }
